pyproject.toml
[tool.poetry]
name = "kvdroid"
version = "0.6.2"
description ="A re-implementation of android java API in python with easy access to some Android functionalitylike Notification,Reading of Contacts, accessing Webview Cookies, etc..."
authors = ["Yunus Ceyhan <yunus.ceyhn@gmail.com>", "Kenechukwu Akubue <kengoon19@gmail.com>"]
readme = "README.md"
keywords = ['Android', 'Androidx', 'Python', 'Kivy', 'KivyMD', 'KvDroid']
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
pyjnius = "^1.6.1"


[tool.poetry.group.dev.dependencies]
setuptools = "^77.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
